+++ merlin/table/tensor_table.py
@@ -15,8 +15,8 @@
 #
 from typing import Any, Dict
 
-from merlin.core.dispatch import df_from_tensor_table, tensor_table_from_df
 from merlin.dag.utils import group_values_offsets
+from merlin.table.conversions import df_from_tensor_table, tensor_table_from_df
 from merlin.table.cupy_column import CupyColumn
 from merlin.table.numpy_column import NumpyColumn
 from merlin.table.tensor_column import TensorColumn, create_tensor_column
